get junction annotations from gene model
